+++ pkg/catalog/types.go
@@ -109,7 +109,7 @@ const (
 const (
 	System_User    = uint32(0)
 	System_Role    = uint32(0)
-	System_Account = uint32(0)
+	System_Account = int64(0)
 )
 
 const (
